requirements.txt
Django<=4.2
djangorestframework
djongo==1.3.6
drf-yasg
pymongo==3.12.0
gunicorn